backend/requirements.txt
aiohttp==3.10.10
asgiref==3.7.2
bencodepy==0.9.5
cn2an==0.5.22
Django==4.2.16
djangorestframework==3.14.0
django-environ==0.10.0
django-filter==23.5
django-polymorphic==3.1.0
lxml==4.9.4
qbittorrent-api==2023.11.57
requests==2.31.0
tzlocal==4.2
whitenoise==6.2.0
apscheduler==3.10.4
openai==1.57.0
psutil==5.9.2
packaging==23.2